feat: add an iteration upgrading the codebase and its structure
