+++ pyproject.toml
@@ -1,0 +1,13 @@
+[tool.black]
+line-length = 120
+target-version = ['py313']
+skip-string-normalization = false
+exclude = '''
+(
+    \.git
+  | \.venv
+  | build
+  | dist
+  | src/__main__.py
+)
+'''
