+++ requirements.txt
@@ -1,0 +1,7 @@
+black==25.1.0
+click==8.2.1
+colorama==0.4.6
+mypy_extensions==1.1.0
+packaging==25.0
+pathspec==0.12.1
+platformdirs==4.3.8
